price handler helper
