+++ modules/token/lib/coingecko-data.service.ts
@@ -99,137 +99,6 @@ export class CoingeckoDataService {
         this.apiKeyParam = env.COINGECKO_API_KEY ? `&x_cg_pro_api_key=${env.COINGECKO_API_KEY}` : '';
     }
 
-    public async updatePricesForTokensWithCoingeckoIds(tokensWithCoingeckoIds: PrismaTokenWithTypes[]) {
-        const timestamp = timestampRoundedUpToNearestHour();
-        const timestampMidnight = timestampEndOfDayMidnight();
-        const updated: PrismaTokenWithTypes[] = [];
-
-        const uniqueTokensWithIds = _.uniqBy(tokensWithCoingeckoIds, 'coingeckoTokenId');
-
-        const chunks = _.chunk(uniqueTokensWithIds, 250); //max page size is 250
-
-        for (const chunk of chunks) {
-            const response = await this.getMarketDataForTokenIds(chunk.map((item) => item.coingeckoTokenId || ''));
-            let operations: any[] = [];
-
-            for (const item of response) {
-                const tokensToUpdate = tokensWithCoingeckoIds.filter((token) => token.coingeckoTokenId === item.id);
-                for (const tokenToUpdate of tokensToUpdate) {
-                    // only update if we have a new price and if we have a price at all
-                    if (moment(item.last_updated).isAfter(moment().subtract(10, 'minutes')) && item.current_price) {
-                        const data = {
-                            price: item.current_price,
-                            ath: item.ath ?? undefined,
-                            atl: item.atl ?? undefined,
-                            marketCap: item.market_cap ?? undefined,
-                            fdv: item.fully_diluted_valuation ?? undefined,
-                            high24h: item.high_24h ?? undefined,
-                            low24h: item.low_24h ?? undefined,
-                            priceChange24h: item.price_change_24h ?? undefined,
-                            priceChangePercent24h: item.price_change_percentage_24h ?? undefined,
-                            priceChangePercent7d: item.price_change_percentage_7d_in_currency ?? undefined,
-                            priceChangePercent14d: item.price_change_percentage_14d_in_currency ?? undefined,
-                            priceChangePercent30d: item.price_change_percentage_30d_in_currency ?? undefined,
-                            updatedAt: item.last_updated,
-                        };
-
-                        operations.push(
-                            prisma.prismaTokenDynamicData.upsert({
-                                where: {
-                                    tokenAddress_chain: {
-                                        tokenAddress: tokenToUpdate.address,
-                                        chain: tokenToUpdate.chain,
-                                    },
-                                },
-                                update: data,
-                                create: {
-                                    coingeckoId: item.id,
-                                    tokenAddress: tokenToUpdate.address,
-                                    chain: tokenToUpdate.chain,
-                                    ...data,
-                                },
-                            }),
-                        );
-
-                        // upsert hourly price and price data, for every chain
-                        operations.push(
-                            prisma.prismaTokenPrice.upsert({
-                                where: {
-                                    tokenAddress_timestamp_chain: {
-                                        tokenAddress: tokenToUpdate.address,
-                                        timestamp,
-                                        chain: tokenToUpdate.chain,
-                                    },
-                                },
-                                update: { price: item.current_price, close: item.current_price },
-                                create: {
-                                    tokenAddress: tokenToUpdate.address,
-                                    chain: tokenToUpdate.chain,
-                                    timestamp,
-                                    price: item.current_price,
-                                    high: item.current_price,
-                                    low: item.current_price,
-                                    open: item.current_price,
-                                    close: item.current_price,
-                                    coingecko: true,
-                                },
-                            }),
-                        );
-
-                        // upsert daily price and price data, for every chain
-                        operations.push(
-                            prisma.prismaTokenPrice.upsert({
-                                where: {
-                                    tokenAddress_timestamp_chain: {
-                                        tokenAddress: tokenToUpdate.address,
-                                        timestamp: timestampMidnight,
-                                        chain: tokenToUpdate.chain,
-                                    },
-                                },
-                                update: { price: item.current_price, close: item.current_price },
-                                create: {
-                                    tokenAddress: tokenToUpdate.address,
-                                    chain: tokenToUpdate.chain,
-                                    timestamp: timestampMidnight,
-                                    price: item.current_price,
-                                    high: item.current_price,
-                                    low: item.current_price,
-                                    open: item.current_price,
-                                    close: item.current_price,
-                                    coingecko: true,
-                                },
-                            }),
-                        );
-
-                        // upsert current price and price data, for every chain
-                        operations.push(
-                            prisma.prismaTokenCurrentPrice.upsert({
-                                where: {
-                                    tokenAddress_chain: {
-                                        tokenAddress: tokenToUpdate.address,
-                                        chain: tokenToUpdate.chain,
-                                    },
-                                },
-                                update: { price: item.current_price },
-                                create: {
-                                    tokenAddress: tokenToUpdate.address,
-                                    chain: tokenToUpdate.chain,
-                                    timestamp,
-                                    price: item.current_price,
-                                    coingecko: true,
-                                },
-                            }),
-                        );
-                        updated.push(tokenToUpdate);
-                    }
-                }
-            }
-
-            await prismaBulkExecuteOperations(operations);
-        }
-        return updated;
-    }
-
     public async syncCoingeckoIds() {
         const allTokens = await prisma.prismaToken.findMany({ where: { chain: networkContext.chain } });
 
@@ -256,124 +125,6 @@ export class CoingeckoDataService {
         }
     }
 
-    // public async initChartData(tokenAddress: string) {
-    //     const latestTimestamp = timestampRoundedUpToNearestHour();
-    //     tokenAddress = tokenAddress.toLowerCase();
-
-    //     const operations: any[] = [];
-    //     const token = await prisma.prismaToken.findUnique({
-    //         where: {
-    //             address_chain: {
-    //                 address: tokenAddress,
-    //                 chain: networkContext.chain,
-    //             },
-    //         },
-    //     });
-
-    //     if (!token || !token.coingeckoTokenId) {
-    //         throw new Error('Missing token or token is missing coingecko token id');
-    //     }
-
-    //     const monthData = await this.getCoinCandlestickData(token.coingeckoTokenId, 30);
-    //     const twentyFourHourData = await this.getCoinCandlestickData(token.coingeckoTokenId, 1);
-
-    //     //merge 30 min data into hourly data
-    //     const hourlyData = Object.values(
-    //         _.groupBy(twentyFourHourData, (item) => timestampRoundedUpToNearestHour(moment.unix(item[0] / 1000))),
-    //     ).map((hourData) => {
-    //         if (hourData.length === 1) {
-    //             const item = hourData[0];
-    //             item[0] = timestampRoundedUpToNearestHour(moment.unix(item[0] / 1000)) * 1000;
-
-    //             return item;
-    //         }
-
-    //         const thirty = hourData[0];
-    //         const hour = hourData[1];
-
-    //         return [hour[0], thirty[1], Math.max(thirty[2], hour[2]), Math.min(thirty[3], hour[3]), hour[4]];
-    //     });
-
-    //     operations.push(prisma.prismaTokenPrice.deleteMany({ where: { tokenAddress, chain: networkContext.chain } }));
-
-    //     operations.push(
-    //         prisma.prismaTokenPrice.createMany({
-    //             data: monthData
-    //                 .filter((item) => item[0] / 1000 <= latestTimestamp)
-    //                 .map((item) => ({
-    //                     tokenAddress,
-    //                     chain: networkContext.chain,
-    //                     timestamp: item[0] / 1000,
-    //                     open: item[1],
-    //                     high: item[2],
-    //                     low: item[3],
-    //                     close: item[4],
-    //                     price: item[4],
-    //                     coingecko: true,
-    //                 })),
-    //         }),
-    //     );
-
-    //     operations.push(
-    //         prisma.prismaTokenPrice.createMany({
-    //             data: hourlyData.map((item) => ({
-    //                 tokenAddress,
-    //                 chain: networkContext.chain,
-    //                 timestamp: Math.floor(item[0] / 1000),
-    //                 open: item[1],
-    //                 high: item[2],
-    //                 low: item[3],
-    //                 close: item[4],
-    //                 price: item[4],
-    //                 coingecko: true,
-    //             })),
-    //             skipDuplicates: true,
-    //         }),
-    //     );
-
-    //     await prismaBulkExecuteOperations(operations, true);
-    // }
-
-    // /**
-    //  *  Rate limit for the CoinGecko API is 10 calls each second per IP address.
-    //  */
-    // public async getTokenPrices(addresses: string[]): Promise<TokenPrices> {
-    //     const addressesPerRequest = addressChunkSize;
-    //     try {
-    //         // if (addresses.length / addressesPerRequest > tokensPerMinute)
-    //         //     throw new Error('Too many requests for rate limit.');
-
-    //         const tokenDefinitions = await tokenService.getTokenDefinitions([networkContext.chain]);
-    //         const mapped = addresses.map((address) => this.getMappedTokenDetails(address, tokenDefinitions));
-    //         const groupedByPlatform = _.groupBy(mapped, 'platform');
-
-    //         const requests: Promise<CoingeckoPriceResponse>[] = [];
-
-    //         _.forEach(groupedByPlatform, (tokens, platform) => {
-    //             const mappedAddresses = tokens.map((token) => token.address);
-    //             const pageCount = Math.ceil(mappedAddresses.length / addressesPerRequest);
-    //             const pages = Array.from(Array(pageCount).keys());
-
-    //             pages.forEach((page) => {
-    //                 const addressString = mappedAddresses.slice(
-    //                     addressesPerRequest * page,
-    //                     addressesPerRequest * (page + 1),
-    //                 );
-    //                 const endpoint = `/simple/token_price/${platform}?contract_addresses=${addressString}&vs_currencies=${this.fiatParam}`;
-    //                 const request = this.get<CoingeckoPriceResponse>(endpoint);
-    //                 requests.push(request);
-    //             });
-    //         });
-
-    //         const paginatedResults = await Promise.all(requests);
-    //         const results = this.parsePaginatedTokens(paginatedResults, mapped);
-
-    //         return results;
-    //     } catch (error: any) {
-    //         throw new Error(`Unable to fetch token prices - ${error.message} - ${error.statusCode}`);
-    //     }
-    // }
-
     // public async getTokenHistoricalPrices(address: string, days: number): Promise<HistoricalPrice[]> {
     //     const now = Math.floor(Date.now() / 1000);
     //     const end = now;
@@ -396,62 +147,11 @@ export class CoingeckoDataService {
     //     }));
     // }
 
-    // private parsePaginatedTokens(paginatedResults: TokenPrices[], mappedTokens: MappedToken[]): TokenPrices {
-    //     const results = paginatedResults.reduce((result, page) => ({ ...result, ...page }), {});
-    //     const prices: TokenPrices = _.mapKeys(results, (val, address) => address);
-
-    //     const resultAddresses = Object.keys(results);
-    //     for (const mappedToken of mappedTokens) {
-    //         if (mappedToken.originalAddress) {
-    //             const resultAddress = resultAddresses.find(
-    //                 (address) => address.toLowerCase() === mappedToken.address.toLowerCase(),
-    //             );
-    //             if (!resultAddress) {
-    //                 console.warn(`Matching address for original address ${mappedToken.originalAddress} not found`);
-    //             } else {
-    //                 prices[mappedToken.originalAddress] = results[resultAddress];
-    //             }
-    //         }
-    //     }
-
-    //     return prices;
-    // }
-
-    /**
-     * Support instances where a token address is not supported by the platform id, provide the option to use a different platform
-    //  */
-    // public getMappedTokenDetails(address: string, tokens: TokenDefinition[]): MappedToken {
-    //     const token = tokens.find((token) => token.address.toLowerCase() === address.toLowerCase());
-    //     if (token && token.coingeckoPlatformId && token.coingeckoContractAddress) {
-    //         return {
-    //             platform: token.coingeckoPlatformId,
-    //             address: isAddress(token.coingeckoContractAddress)
-    //                 ? token.coingeckoContractAddress.toLowerCase()
-    //                 : token.coingeckoContractAddress,
-    //             originalAddress: address.toLowerCase(),
-    //         };
-    //     }
-
-    //     return {
-    //         platform: networkContext.data.coingecko.platformId,
-    //         address: address.toLowerCase(),
-    //     };
-    // }
-
-    private async getMarketDataForTokenIds(tokenIds: string[]): Promise<CoingeckoTokenMarketData[]> {
+    public async getMarketDataForTokenIds(tokenIds: string[]): Promise<CoingeckoTokenMarketData[]> {
         const endpoint = `/coins/markets?vs_currency=${this.fiatParam}&ids=${tokenIds}&per_page=250&page=1&sparkline=false&price_change_percentage=1h%2C24h%2C7d%2C14d%2C30d`;
 
         return this.get<CoingeckoTokenMarketData[]>(endpoint);
     }
-
-    // public async getCoinCandlestickData(
-    //     tokenId: string,
-    //     days: 1 | 30,
-    // ): Promise<[number, number, number, number, number][]> {
-    //     const endpoint = `/coins/${tokenId}/ohlc?vs_currency=usd&days=${days}`;
-
-    //     return this.get(endpoint);
-    // }
 
     private async getCoinIdList(): Promise<CoinId[]> {
         const endpoint = `/coins/list?include_platform=true`;
