+++ modules/token/token.prisma
@@ -55,6 +55,7 @@ model PrismaTokenPrice {
     token               PrismaToken         @relation(fields:[tokenAddress, chain], references: [address, chain])
     chain               Chain               
     updatedAt           DateTime            @updatedAt
+    updatedBy           String?
     timestamp           Int
     price               Float
     coingecko           Boolean?
