+++ prisma/schema.prisma
@@ -633,6 +633,7 @@ model PrismaTokenPrice {
     token               PrismaToken         @relation(fields:[tokenAddress, chain], references: [address, chain])
     chain               Chain               
     updatedAt           DateTime            @updatedAt
+    updatedBy           String?
     timestamp           Int
     price               Float
     coingecko           Boolean?
